requirements.txt
openai
streamlit
transformers
torch
sentence-transformers
faiss-cpu
python-dotenv
